requirements.txt
amqp==2.5.2
asgiref==3.2.5
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.9.0
billiard==3.6.3.0
boto3==1.12.43
botocore==1.15.43
bs4==0.0.1
celery==4.4.2
cent==3.0.1
certifi==2019.11.28
chardet==3.0.4
coverage==5.0.4
dj-database-url==0.5.0
Django==2.0
django-cors-headers==3.2.1
django-cprofile-middleware==1.0.5
django-heroku==0.3.1
django-ranged-response==0.2.0
django-rest-framework==0.1.0
django-simple-captcha==0.5.12
djangorestframework==3.11.0
docutils==0.15.2
factory-boy==2.12.0
Faker==4.0.2
flower==0.9.4
gprof2dot==2019.11.30
gunicorn==20.0.4
humanize==0.5.1
idna==2.9
importlib-metadata==1.5.0
insultgenerator==1.0.3
isort==4.3.21
jmespath==0.9.5
kombu==4.6.8
lazy-object-proxy==1.4.3
loremipsum==1.0.5
mccabe==0.6.1
mock==4.0.2
more-itertools==8.2.0
packaging==20.3
Pillow==7.0.0
pluggy==0.13.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.1
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.4.0
python-dateutil==2.8.1
python-magic==0.4.15
python-memcached==1.59
pytz==2019.3
pywatchman==1.4.1
redis==3.4.1
requests==2.23.0
s3transfer==0.3.3
selenium==3.141.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.3
tornado==6.0.4
typed-ast==1.4.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
whitenoise==5.0.1
wrapt==1.11.0
zipp==3.1.0
